+++ learn2learn/data/meta_dataset.pyx
@@ -1,4 +1,7 @@
+# cython: language_version=3
 #!/usr/bin/env python3
+
+cimport cython
 
 import os
 import pickle